requirements.txt
alembic==1.8.1
autopep8==1.6.0
Babel==2.10.3
blinker==1.5
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dnspython==2.2.1
dominate==2.7.0
email-validator==1.2.1
Flask==2.1.3
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
langdetect==1.0.9
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
pycodestyle==2.8.0
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
toml==0.10.2
urllib3==1.26.11
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
